api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
supabase==2.8.1
websockets>=15.0.1
python-dotenv==1.0.0


